+++ lib/pbio/platform/essential_hub/pbsysconfig.h
@@ -1,8 +1,9 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2022 The Pybricks Authors
+// Copyright (c) 2021-2022 The Pybricks Authors
 
 #define PBSYS_CONFIG_BATTERY_CHARGER                (1)
 #define PBSYS_CONFIG_BLUETOOTH                      (1)
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (0)
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_STATUS_LIGHT                   (1)
+#define PBSYS_CONFIG_STATUS_LIGHT_BATTERY           (1)