+++ lib/pbio/platform/nxt/pbsysconfig.h
@@ -6,3 +6,4 @@
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (0)
 #define PBSYS_CONFIG_MAIN                           (1)
 #define PBSYS_CONFIG_STATUS_LIGHT                   (0)
+#define PBSYS_CONFIG_STATUS_LIGHT_BATTERY           (0)